+++ tests/data/WritingToReadonlyAttribute/NonAttributeClass.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Icanhazstring\PhpstanReadonlyPropertyExtension\Test\data;
+namespace Icanhazstring\PhpstanReadonlyPropertyExtension\Test\data\WritingToReadonlyAttribute;
 
 final class NonAttributeClass
 {